Just changed the file locations for the main page components.
