+++ components/common/cards/player/PlayerCard.jsx
@@ -12,72 +12,66 @@ If need be, this can just be put into MainPage.js.
 on press it 
 */
 
-const PlayerCard = ({item}, handleNavigate) => {
-  
+const PlayerCard = ({ item }, handleNavigate) => {
   // console.log(item)
   // may need to change in the future if doesn't catch all errors
-  const imagePath = (item.playerHeadshot) ? `../../../../assets/images/playerHeadShots${item.playerHeadshot}` : "../../../../assets/images/playerHeadShots/defaultPlayer.jpeg"
+  const imagePath = item.playerHeadshot
+    ? `../../../../assets/images/playerHeadShots${item.playerHeadshot}`
+    : "../../../../assets/images/playerHeadShots/defaultPlayer.jpeg";
 
   let image = "";
-  
-  if(!item.playerHeadshot){
+
+  if (!item.playerHeadshot) {
     image = require("../../../../assets/images/playerHeadShots/defaultPlayer.jpeg");
 
-    return(
+    return (
       // [playerFile, setPlayerFile] = useState();
-    
+
       // <Link href="/testpage" asChild>
-        <TouchableOpacity onPress={(handleNavigate)}>
-          <Card style={style.cardContainer}>
-            <Card.Cover
-              style={style.imageLayout}
-              source={image}
-            />
-            <Card.Cover style={style.statsLayout} />
-            <Card.Content>
-              <Text style={style.nameLayout}>{item.playerName}</Text>
-              <Text style={style.teamLayout}>{item.teamAbv}</Text>
-              <Text style={style.statsTitle}>Stats</Text>
-              <Text style={style.statsList}>
-                PPG: {item.PTS}
-                AST: {item.AST}
-                REB: {item.REB}
-                FG%: {item.FG_PERCENT}
-              </Text>
-            </Card.Content>
-          </Card>
-        </TouchableOpacity>
+      <TouchableOpacity onPress={handleNavigate}>
+        <Card style={style.cardContainer}>
+          <Card.Cover style={style.imageLayout} source={image} />
+          <Card.Cover style={style.statsLayout} />
+          <Card.Content>
+            <Text style={style.nameLayout}>{item.playerName}</Text>
+            <Text style={style.teamLayout}>{item.teamAbv}</Text>
+            <Text style={style.statsTitle}>Stats</Text>
+            <Text style={style.statsList}>
+              PPG: {item.PTS}
+              AST: {item.AST}
+              REB: {item.REB}
+              FG%: {item.FG_PERCENT}
+            </Text>
+          </Card.Content>
+        </Card>
+      </TouchableOpacity>
       // </Link>
-    )
+    );
+  } else {
+    return (
+      <TouchableOpacity onPress={handleNavigate}>
+        <Card style={style.cardContainer}>
+          <Card.Cover
+            style={style.imageLayout}
+            source={{ uri: item.playerHeadshot }}
+          />
+          <Card.Cover style={style.statsLayout} />
+          <Card.Content>
+            <Text style={style.nameLayout}>{item.playerName}</Text>
+            <Text style={style.teamLayout}>{item.teamAbv}</Text>
+            <Text style={style.statsTitle}>Stats</Text>
+            <Text style={style.statsList}>
+              PPG: {item.PTS} {"\n"}
+              AST: {item.AST} {"\n"}
+              REB: {item.REB} {"\n"}
+              FG%: {item.FG_PERCENT} {"\n"}
+            </Text>
+          </Card.Content>
+        </Card>
+      </TouchableOpacity>
+    );
   }
-  else{
-    return(
-    <TouchableOpacity onPress={(handleNavigate)}>
-          <Card style={style.cardContainer}>
-            <Card.Cover
-              style={style.imageLayout}
-              source={{uri: item.playerHeadshot}}
-            />
-            <Card.Cover style={style.statsLayout} />
-            <Card.Content>
-              <Text style={style.nameLayout}>{item.playerName}</Text>
-              <Text style={style.teamLayout}>{item.teamAbv}</Text>
-              <Text style={style.statsTitle}>Stats</Text>
-              <Text style={style.statsList}>
-                PPG: {item.PTS}
-                AST: {item.AST}
-                REB: {item.REB}
-                FG%: {item.FG_PERCENT}
-              </Text>
-            </Card.Content>
-          </Card>
-        </TouchableOpacity>
-    )
-  }
-  
-
-
-  };
+};
 
 export default PlayerCard;
 
@@ -85,15 +79,15 @@ const style = StyleSheet.create({
   cardContainer: {
     backgroundColor: COLORS.orange,
     margin: 5,
-    width: 140,
-    height: 120,
+    width: 160,
+    height: 125,
   },
   imageLayout: {
     marginTop: 5,
     marginHorizontal: 3,
-    width: 80,
-    height: 80,
-    borderWidth: 0.1,
+    width: 85,
+    height: 85,
+    borderWidth: 0.5,
     borderRadius: 10,
     borderColor: "#121212",
     backgroundColor: "#121212",
@@ -104,6 +98,7 @@ const style = StyleSheet.create({
     fontSize: 10,
     position: "absolute",
     paddingLeft: 3,
+    fontWeight: "bold",
   },
   teamLayout: {
     textAlign: "left",
@@ -119,17 +114,17 @@ const style = StyleSheet.create({
     position: "absolute",
     backgroundColor: COLORS.dark,
     right: 0,
-    height: 110,
-    width: 50,
+    height: 115,
+    width: 65,
   },
   statsList: {
     textAlign: "left",
     color: COLORS.light,
     fontSize: 10,
     position: "absolute",
-    right: 10,
-    top: -60,
-    width: 40,
+    right: 9,
+    top: -65,
+    width: 55,
   },
   statsTitle: {
     textAlign: "center",
@@ -137,7 +132,8 @@ const style = StyleSheet.create({
     fontSize: 10,
     position: "absolute",
     right: 10,
-    top: -80,
-    width: 40,
+    top: -85,
+    width: 50,
+    fontWeight: "bold",
   },
 });
